+++ internal/write/service_test.go
@@ -7,6 +7,7 @@ package write
 import (
 	"context"
 	"errors"
+	"fmt"
 	"strings"
 	"sync"
 	"testing"
@@ -19,6 +20,15 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func TestPrecisionToString(t *testing.T) {
+	assert.Equal(t, "ns", precisionToString(time.Nanosecond))
+	assert.Equal(t, "us", precisionToString(time.Microsecond))
+	assert.Equal(t, "ms", precisionToString(time.Millisecond))
+	assert.Equal(t, "s", precisionToString(time.Second))
+	assert.Equal(t, "ns", precisionToString(time.Hour))
+	assert.Equal(t, "ns", precisionToString(time.Microsecond*20))
+}
 
 func TestAddDefaultTags(t *testing.T) {
 	hs := test.NewTestService(t, "http://localhost:8888")
@@ -67,77 +77,99 @@ func TestRetryStrategy(t *testing.T) {
 	opts := write.DefaultOptions().SetRetryInterval(1)
 	ctx := context.Background()
 	srv := NewService("my-org", "my-bucket", hs, opts)
+	// Set permanent reply error to force writes fail and retry
 	hs.SetReplyError(&http.Error{
 		StatusCode: 429,
 	})
-	b1 := NewBatch("1\n", opts.RetryInterval())
+	// This batch will fail and it be added to retry queue
+	b1 := NewBatch("1\n", opts.RetryInterval(), opts.MaxRetryTime())
 	err := srv.HandleWrite(ctx, b1)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(1), b1.retryDelay)
+	assert.EqualValues(t, 1, b1.retryDelay)
 	assert.Equal(t, 1, srv.retryQueue.list.Len())
 
 	//wait retry delay + little more
 	<-time.After(time.Millisecond*time.Duration(b1.retryDelay) + time.Microsecond*5)
-	b2 := NewBatch("2\n", opts.RetryInterval())
+	// First batch will be tried to write again and this one will added to retry queue
+	b2 := NewBatch("2\n", opts.RetryInterval(), opts.MaxRetryTime())
 	err = srv.HandleWrite(ctx, b2)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(5), b1.retryDelay)
+	assertBetween(t, b1.retryDelay, 2, 4)
 	assert.Equal(t, 2, srv.retryQueue.list.Len())
 
 	//wait retry delay + little more
 	<-time.After(time.Millisecond*time.Duration(b1.retryDelay) + time.Microsecond*5)
-	b3 := NewBatch("3\n", opts.RetryInterval())
+	// First batch will be tried to write again and this one will added to retry queue
+	b3 := NewBatch("3\n", opts.RetryInterval(), opts.MaxRetryTime())
 	err = srv.HandleWrite(ctx, b3)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(25), b1.retryDelay)
+	assertBetween(t, b1.retryDelay, 4, 8)
 	assert.Equal(t, 3, srv.retryQueue.list.Len())
 
-	// let write pass and it will clear queue
+	//wait retry delay + little more
 	<-time.After(time.Millisecond*time.Duration(b1.retryDelay) + time.Microsecond*5)
+	// First batch will be tried to write again and this one will added to retry queue
+	b4 := NewBatch("4\n", opts.RetryInterval(), opts.MaxRetryTime())
+	err = srv.HandleWrite(ctx, b4)
+	assert.NotNil(t, err)
+	assertBetween(t, b1.retryDelay, 8, 16)
+	assert.Equal(t, 4, srv.retryQueue.list.Len())
+
+	<-time.After(time.Millisecond*time.Duration(b1.retryDelay) + time.Microsecond*5)
+	// Clear error and let write pass
 	hs.SetReplyError(nil)
-	err = srv.HandleWrite(ctx, NewBatch("4\n", opts.RetryInterval()))
+	// Batches from retry queue will be sent first
+	err = srv.HandleWrite(ctx, NewBatch("5\n", opts.RetryInterval(), opts.MaxRetryTime()))
 	assert.Nil(t, err)
 	assert.Equal(t, 0, srv.retryQueue.list.Len())
-	require.Len(t, hs.Lines(), 4)
+	require.Len(t, hs.Lines(), 5)
 	assert.Equal(t, "1", hs.Lines()[0])
 	assert.Equal(t, "2", hs.Lines()[1])
 	assert.Equal(t, "3", hs.Lines()[2])
 	assert.Equal(t, "4", hs.Lines()[3])
+	assert.Equal(t, "5", hs.Lines()[4])
 }
 
 func TestBufferOverwrite(t *testing.T) {
 	log.Log.SetLogLevel(log.DebugLevel)
 	hs := test.NewTestService(t, "http://localhost:8086")
-	// Buffer limit 15000, bach ii 5000 => buffer for 3 batches
+	// Buffer limit 15000, bach is 5000 => buffer for 3 batches
 	opts := write.DefaultOptions().SetRetryInterval(1).SetRetryBufferLimit(15000)
 	ctx := context.Background()
 	srv := NewService("my-org", "my-bucket", hs, opts)
+	// Set permanent reply error to force writes fail and retry
 	hs.SetReplyError(&http.Error{
 		StatusCode: 429,
 	})
-	b1 := NewBatch("1\n", opts.RetryInterval())
+	// This batch will fail and it be added to retry queue
+	b1 := NewBatch("1\n", opts.RetryInterval(), opts.MaxRetryTime())
 	err := srv.HandleWrite(ctx, b1)
 	assert.NotNil(t, err)
 	assert.Equal(t, uint(1), b1.retryDelay)
 	assert.Equal(t, 1, srv.retryQueue.list.Len())
 
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
-	b2 := NewBatch("2\n", opts.RetryInterval())
+	b2 := NewBatch("2\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// First batch will be tried to write again and this one will added to retry queue
 	err = srv.HandleWrite(ctx, b2)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(5), b1.retryDelay)
+	assertBetween(t, b1.retryDelay, 2, 4)
 	assert.Equal(t, 2, srv.retryQueue.list.Len())
 
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
-	b3 := NewBatch("3\n", opts.RetryInterval())
+	b3 := NewBatch("3\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// First batch will be tried to write again and this one will added to retry queue
 	err = srv.HandleWrite(ctx, b3)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(25), b1.retryDelay)
+	assertBetween(t, b1.retryDelay, 4, 8)
 	assert.Equal(t, 3, srv.retryQueue.list.Len())
 
-	// Write early
-	<-time.After(time.Millisecond * time.Duration(b1.retryDelay) / 2)
-	b4 := NewBatch("4\n", opts.RetryInterval())
+	// Write early and overwrite
+	<-time.After(time.Millisecond)
+	b4 := NewBatch("4\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// No write will occur, because retry delay has not passed yet
+	// However new bach will be added to retry queue. Retry queue has limit 3,
+	// so first batch will be discarded
 	err = srv.HandleWrite(ctx, b4)
 	assert.NoError(t, err)
 	assert.Equal(t, uint(1), b2.retryDelay)
@@ -145,16 +177,20 @@ func TestBufferOverwrite(t *testing.T) {
 
 	// Overwrite
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay) / 2)
-	b5 := NewBatch("5\n", opts.RetryInterval())
+	b5 := NewBatch("5\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// Second batch will be tried to write again
+	// However, write will fail and as new batch is added to retry queue
+	// the second batch will be discarded
 	err = srv.HandleWrite(ctx, b5)
 	assert.Error(t, err)
-	assert.Equal(t, uint(5), b2.retryDelay)
+	assertBetween(t, b2.retryDelay, 2, 4)
 	assert.Equal(t, 3, srv.retryQueue.list.Len())
 
-	// let write pass and it will clear queue
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
+	// Clear error and let write pass
 	hs.SetReplyError(nil)
-	err = srv.HandleWrite(ctx, NewBatch("6\n", opts.RetryInterval()))
+	// Batches from retry queue will be sent first
+	err = srv.HandleWrite(ctx, NewBatch("6\n", opts.RetryInterval(), opts.MaxRetryTime()))
 	assert.Nil(t, err)
 	assert.Equal(t, 0, srv.retryQueue.list.Len())
 	require.Len(t, hs.Lines(), 4)
@@ -167,32 +203,130 @@ func TestBufferOverwrite(t *testing.T) {
 func TestMaxRetryInterval(t *testing.T) {
 	log.Log.SetLogLevel(log.DebugLevel)
 	hs := test.NewTestService(t, "http://localhost:8086")
-	//
-	opts := write.DefaultOptions().SetRetryInterval(1).SetMaxRetryInterval(10)
+	// MaxRetryInterval only 4ms, will be reached quickly
+	opts := write.DefaultOptions().SetRetryInterval(1).SetMaxRetryInterval(4)
 	ctx := context.Background()
 	srv := NewService("my-org", "my-bucket", hs, opts)
+	// Set permanent reply error to force writes fail and retry
 	hs.SetReplyError(&http.Error{
 		StatusCode: 503,
 	})
-	b1 := NewBatch("1\n", opts.RetryInterval())
+	// This batch will fail and it be added to retry queue
+	b1 := NewBatch("1\n", opts.RetryInterval(), opts.MaxRetryTime())
 	err := srv.HandleWrite(ctx, b1)
 	assert.NotNil(t, err)
 	assert.Equal(t, uint(1), b1.retryDelay)
 	assert.Equal(t, 1, srv.retryQueue.list.Len())
 
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
-	b2 := NewBatch("2\n", opts.RetryInterval())
+	b2 := NewBatch("2\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// First batch will be tried to write again and this one will added to retry queue
 	err = srv.HandleWrite(ctx, b2)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(5), b1.retryDelay)
+	assertBetween(t, b1.retryDelay, 2, 4)
 	assert.Equal(t, 2, srv.retryQueue.list.Len())
 
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
-	b3 := NewBatch("3\n", opts.RetryInterval())
+	b3 := NewBatch("3\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// First batch will be tried to write again and this one will added to retry queue
 	err = srv.HandleWrite(ctx, b3)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(10), b1.retryDelay)
+	// New computed delay of first batch should be 4-8, is limited to 4
+	assert.EqualValues(t, 4, b1.retryDelay)
 	assert.Equal(t, 3, srv.retryQueue.list.Len())
+
+	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
+	b4 := NewBatch("4\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// First batch will be tried to write again and this one will added to retry queue
+	err = srv.HandleWrite(ctx, b4)
+	assert.NotNil(t, err)
+	// New computed delay of first batch should be 8-116, is limited to 4
+	assert.EqualValues(t, 4, b1.retryDelay)
+	assert.Equal(t, 4, srv.retryQueue.list.Len())
+}
+
+func TestMaxRetries(t *testing.T) {
+	log.Log.SetLogLevel(log.DebugLevel)
+	hs := test.NewTestService(t, "http://localhost:8086")
+	opts := write.DefaultOptions().SetRetryInterval(1)
+	ctx := context.Background()
+	srv := NewService("my-org", "my-bucket", hs, opts)
+	// Set permanent reply error to force writes fail and retry
+	hs.SetReplyError(&http.Error{
+		StatusCode: 429,
+	})
+	// This batch will fail and it be added to retry queue
+	b1 := NewBatch("1\n", opts.RetryInterval(), opts.MaxRetryTime())
+	err := srv.HandleWrite(ctx, b1)
+	assert.NotNil(t, err)
+	assert.EqualValues(t, 1, b1.retryDelay)
+	assert.Equal(t, 1, srv.retryQueue.list.Len())
+	// Write so many batches as it is maxRetries (5)
+	// First batch will be written and it will reach max retry limit
+	for i, e := uint(1), uint(2); i <= opts.MaxRetries(); i++ {
+		//wait retry delay + little more
+		<-time.After(time.Millisecond*time.Duration(b1.retryDelay) + time.Microsecond*5)
+		b := NewBatch(fmt.Sprintf("%d\n", i+1), opts.RetryInterval(), opts.MaxRetryTime())
+		err = srv.HandleWrite(ctx, b)
+		assert.NotNil(t, err)
+		assertBetween(t, b1.retryDelay, e, e*2)
+		exp := min(i+1, opts.MaxRetries())
+		assert.EqualValues(t, exp, srv.retryQueue.list.Len())
+		e *= 2
+	}
+	//Test if was removed from retry queue
+	assert.True(t, b1.evicted)
+
+	<-time.After(time.Millisecond*time.Duration(b1.retryDelay) + time.Microsecond*5)
+	// Clear error and let write pass
+	hs.SetReplyError(nil)
+	// Batches from retry queue will be sent first
+	err = srv.HandleWrite(ctx, NewBatch(fmt.Sprintf("%d\n", opts.MaxRetries()+2), opts.RetryInterval(), opts.MaxRetryTime()))
+	assert.Nil(t, err)
+	assert.Equal(t, 0, srv.retryQueue.list.Len())
+	require.Len(t, hs.Lines(), int(opts.MaxRetries()+1))
+	for i := uint(2); i <= opts.MaxRetries()+2; i++ {
+		assert.Equal(t, fmt.Sprintf("%d", i), hs.Lines()[i-2])
+	}
+}
+
+func TestMaxRetryTime(t *testing.T) {
+	log.Log.SetLogLevel(log.DebugLevel)
+	hs := test.NewTestService(t, "http://localhost:8086")
+	// Set maxRetryTime 5ms
+	opts := write.DefaultOptions().SetRetryInterval(1).SetMaxRetryTime(5)
+	ctx := context.Background()
+	srv := NewService("my-org", "my-bucket", hs, opts)
+	// Set permanent reply error to force writes fail and retry
+	hs.SetReplyError(&http.Error{
+		StatusCode: 429,
+	})
+	// This batch will fail and it be added to retry queue and it will expire 5ms after
+	b1 := NewBatch("1\n", opts.RetryInterval(), opts.MaxRetryTime())
+	err := srv.HandleWrite(ctx, b1)
+	assert.NotNil(t, err)
+	assert.EqualValues(t, 1, b1.retryDelay)
+	assert.Equal(t, 1, srv.retryQueue.list.Len())
+
+	// Wait for batch expiration
+	<-time.After(5 * time.Millisecond)
+	b := NewBatch("2\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// First batch will be tried to write again and it will be checked agains maxRetryTime. New batch will added to retry queue
+	err = srv.HandleWrite(ctx, b)
+	require.NotNil(t, err)
+	// Error about batch expiration
+	assert.Equal(t, "write failed (attempts 1): max retry time exceeded", err.Error())
+	assert.Equal(t, 1, srv.retryQueue.list.Len())
+
+	// Clear error and let write pass
+	hs.SetReplyError(nil)
+	// A batch from retry queue will be sent first
+	err = srv.HandleWrite(ctx, NewBatch("3\n", opts.RetryInterval(), opts.MaxRetryTime()))
+	assert.Nil(t, err)
+	assert.Equal(t, 0, srv.retryQueue.list.Len())
+	require.Len(t, hs.Lines(), 2)
+	assert.Equal(t, "2", hs.Lines()[0])
+	assert.Equal(t, "3", hs.Lines()[1])
 }
 
 func TestRetryOnConnectionError(t *testing.T) {
@@ -203,34 +337,41 @@ func TestRetryOnConnectionError(t *testing.T) {
 	ctx := context.Background()
 	srv := NewService("my-org", "my-bucket", hs, opts)
 
+	// Set permanent non HTTP  error to force writes fail and retry
 	hs.SetReplyError(&http.Error{
 		Err: errors.New("connection refused"),
 	})
 
-	b1 := NewBatch("1\n", opts.RetryInterval())
+	// This batch will fail and it be added to retry queue
+	b1 := NewBatch("1\n", opts.RetryInterval(), opts.MaxRetryTime())
 	err := srv.HandleWrite(ctx, b1)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(1), b1.retryDelay)
+	assert.EqualValues(t, 1, b1.retryDelay)
 	assert.Equal(t, 1, srv.retryQueue.list.Len())
 
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
-	b2 := NewBatch("2\n", opts.RetryInterval())
+
+	b2 := NewBatch("2\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// First batch will be tried to write again and this one will added to retry queue
 	err = srv.HandleWrite(ctx, b2)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(5), b1.retryDelay)
+	assertBetween(t, b1.retryDelay, 2, 4)
 	assert.Equal(t, 2, srv.retryQueue.list.Len())
 
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
-	b3 := NewBatch("3\n", opts.RetryInterval())
+
+	b3 := NewBatch("3\n", opts.RetryInterval(), opts.MaxRetryTime())
+	// First batch will be tried to write again and this one will added to retry queue
 	err = srv.HandleWrite(ctx, b3)
 	assert.NotNil(t, err)
-	assert.Equal(t, uint(25), b1.retryDelay)
+	assertBetween(t, b1.retryDelay, 4, 8)
 	assert.Equal(t, 3, srv.retryQueue.list.Len())
 
-	// let write pass and it will clear queue
 	<-time.After(time.Millisecond * time.Duration(b1.retryDelay))
+	// Clear error and let write pass
 	hs.SetReplyError(nil)
-	err = srv.HandleWrite(ctx, NewBatch("4\n", opts.RetryInterval()))
+	// Batches from retry queue will be sent first
+	err = srv.HandleWrite(ctx, NewBatch("4\n", opts.RetryInterval(), opts.MaxRetryTime()))
 	assert.Nil(t, err)
 	assert.Equal(t, 0, srv.retryQueue.list.Len())
 	require.Len(t, hs.Lines(), 4)
@@ -238,7 +379,6 @@ func TestRetryOnConnectionError(t *testing.T) {
 	assert.Equal(t, "2", hs.Lines()[1])
 	assert.Equal(t, "3", hs.Lines()[2])
 	assert.Equal(t, "4", hs.Lines()[3])
-
 }
 
 func TestNoRetryIfMaxRetriesIsZero(t *testing.T) {
@@ -253,7 +393,7 @@ func TestNoRetryIfMaxRetriesIsZero(t *testing.T) {
 		Err: errors.New("connection refused"),
 	})
 
-	b1 := NewBatch("1\n", opts.RetryInterval())
+	b1 := NewBatch("1\n", opts.RetryInterval(), opts.MaxRetryTime())
 	err := srv.HandleWrite(ctx, b1)
 	assert.NotNil(t, err)
 	assert.Equal(t, 0, srv.retryQueue.list.Len())
@@ -270,11 +410,36 @@ func TestWriteContextCancel(t *testing.T) {
 	wg.Add(1)
 	go func() {
 		<-time.After(10 * time.Millisecond)
-		err = srv.HandleWrite(ctx, NewBatch(strings.Join(lines, "\n"), opts.RetryInterval()))
+		err = srv.HandleWrite(ctx, NewBatch(strings.Join(lines, "\n"), opts.RetryInterval(), opts.MaxRetryTime()))
 		wg.Done()
 	}()
 	cancel()
 	wg.Wait()
 	require.Equal(t, context.Canceled, err)
 	assert.Len(t, hs.Lines(), 0)
+}
+
+func TestPow(t *testing.T) {
+	assert.EqualValues(t, 1, pow(10, 0))
+	assert.EqualValues(t, 10, pow(10, 1))
+	assert.EqualValues(t, 4, pow(2, 2))
+	assert.EqualValues(t, 1, pow(1, 2))
+	assert.EqualValues(t, 125, pow(5, 3))
+}
+
+func assertBetween(t *testing.T, val, min, max uint) {
+	t.Helper()
+	assert.True(t, val >= min && val <= max, fmt.Sprintf("%d is outside <%d;%d>", val, min, max))
+}
+
+func TestComputeRetryDelay(t *testing.T) {
+	hs := test.NewTestService(t, "http://localhost:8888")
+	opts := write.DefaultOptions()
+	srv := NewService("my-org", "my-bucket", hs, opts)
+	assertBetween(t, srv.computeRetryDelay(0), 5_000, 10_000)
+	assertBetween(t, srv.computeRetryDelay(1), 10_000, 20_000)
+	assertBetween(t, srv.computeRetryDelay(2), 20_000, 40_000)
+	assertBetween(t, srv.computeRetryDelay(3), 40_000, 80_000)
+	assertBetween(t, srv.computeRetryDelay(4), 80_000, 125_000)
+	assert.EqualValues(t, 125_000, srv.computeRetryDelay(5))
 }